refactor(xapiri): extract dashboard_term.go — embedded terminal helpers
watchPTYCmd, processTermBytes, keyMsgToBytes, stripNonSGR, termRawToLines,
renderTermPane, and termPane constants form one coherent PTY read/render
pipeline. Moved to dashboard_term.go per ADR 0014.
